Struct sval_json::Writer

pub struct Writer<W>(_);

A stream for writing structured data as json.

The stream internally wraps a std::io::Write.

Examples

Create an owned json stream:

use sval::stream::OwnedStream;
use sval_json::Writer;

let mut stream = OwnedStream::new(Writer::new(Vec::<u8>::new()));
stream.any(42)?;
let json = stream.into_inner().end()?;

assert_eq!(Some("42"), str::from_utf8(&json).ok());

Methods

impl<W> Writer<W> where
W: Write,

pub fn new(out: W) -> Self

Create a new json stream.

pub fn is_valid(&self) -> bool

Whether the stream has seen a complete, valid json structure.

pub fn end(self) -> Result<W, End<Self>>

Complete the stream and return the inner writer.

If the writer contains incomplete json then this method will fail. The returned error can be used to pull the original stream back out.

pub fn into_inner_unchecked(self) -> W

Get the inner writer back out of the stream without ensuring it's valid.